+++ main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main(){
-  int p;
+  long p;
   cout << "HolaMundo!" << endl;
   cout << "Ingresar P para ejercicio 2.  P: ";
   cin >> p;
